fix: move fingerpint package to utils for it to be used by others
