+++ exporter/clickhouselogsexporter/exporter.go
@@ -27,9 +27,9 @@ import (
 
 	"github.com/ClickHouse/clickhouse-go/v2"
 	driver "github.com/ClickHouse/clickhouse-go/v2/lib/driver"
-	"github.com/SigNoz/signoz-otel-collector/exporter/clickhouselogsexporter/logsv2"
 	"github.com/SigNoz/signoz-otel-collector/usage"
 	"github.com/SigNoz/signoz-otel-collector/utils"
+	"github.com/SigNoz/signoz-otel-collector/utils/fingerprint"
 	"github.com/google/uuid"
 	"github.com/segmentio/ksuid"
 	"go.opencensus.io/stats"
@@ -332,7 +332,7 @@ func (e *clickhouseLogsExporter) pushToClickhouse(ctx context.Context, ld plog.L
 					}
 					fp, exists := resourcesSeen[int64(lBucketStart)][resourceJson]
 					if !exists {
-						fp = logsv2.CalculateFingerprint(res.Attributes().AsRaw(), logsv2.ResourceHierarchy())
+						fp = fingerprint.CalculateFingerprint(res.Attributes().AsRaw(), fingerprint.ResourceHierarchy())
 						resourcesSeen[int64(lBucketStart)][resourceJson] = fp
 					}
